refactor: stamp 엔티티의 day 생성시 초기 값 변경 프론트 요구사항에 맞추어 초기값을 totalDays 값만큼 "u"로 넣어두게 설정. 스탬프 요청이 오면 "u"를 "o" 혹은 "x" 로 변경하도록 수정.

+++ src/main/java/challenge/nDaysChallenge/domain/room/Period.java
@@ -15,11 +15,11 @@ public class Period {
 
     private LocalDate startDate;
     private LocalDate endDate;
-    private Long totalDays;
+    private int totalDays;
 
 
     //==생성 메서드==//
-    public Period(LocalDate startDate, Long totalDays) {
+    public Period(LocalDate startDate, int totalDays) {
         this.startDate = startDate;
         this.endDate = startDate.plusDays(totalDays-1);
         this.totalDays = totalDays;
