+++ src/main/java/challenge/nDaysChallenge/dto/request/Room/RoomRequestDto.java
@@ -13,7 +13,7 @@ public class RoomRequestDto {
     private String type;
     private String name;
     private String category;
-    private Long totalDays;
+    private int totalDays;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate startDate;
     private int passCount;
@@ -24,7 +24,7 @@ public class RoomRequestDto {
 
 
     @Builder
-    public RoomRequestDto(String name, String category, String reward, int passCount, LocalDate startDate, Long totalDays, String type, int successCount, int usedPassCount) {
+    public RoomRequestDto(String name, String category, String reward, int passCount, LocalDate startDate, int totalDays, String type, int successCount, int usedPassCount) {
         this.type = type;
         this.name = name;
         this.category = category;
