+++ src/main/java/challenge/nDaysChallenge/dto/response/Room/GroupRoomResponseDto.java
@@ -15,7 +15,7 @@ public class GroupRoomResponseDto {
     private String type;
     private String name;
     private String category;
-    private Long totalDays;
+    private int totalDays;
     @JsonFormat(pattern = "yyyy-MM-dd")
     private LocalDate startDate;
     @JsonFormat(pattern = "yyyy-MM-dd")
@@ -28,7 +28,7 @@ public class GroupRoomResponseDto {
     private Set<Long> groupMembers = new HashSet<>();
 
     @Builder
-    public GroupRoomResponseDto(Long roomNumber, String name, String category, String reward, int passCount, String type, String status, LocalDate startDate, LocalDate endDate, Long totalDays, Set<Long> groupMembers) {
+    public GroupRoomResponseDto(Long roomNumber, String name, String category, String reward, int passCount, String type, String status, LocalDate startDate, LocalDate endDate, int totalDays, Set<Long> groupMembers) {
         this.roomNumber = roomNumber;
         this.type = type;
         this.name = name;
